+++ src/node/node_state.h
@@ -18,13 +18,13 @@
 #include "history.h"
 #include "indexing/indexer.h"
 #include "js/wrap.h"
-#include "kv/snapshotter.h"
 #include "network_state.h"
 #include "node/attestation_types.h"
 #include "node/hooks.h"
 #include "node/http_node_client.h"
 #include "node/jwt_key_auto_refresh.h"
 #include "node/node_to_node_channel_manager.h"
+#include "node/snapshotter.h"
 #include "node_to_node.h"
 #include "resharing.h"
 #include "rpc/frontend.h"
