+++ src/node/snapshot_serdes.h
