+++ src/node/test/snapshotter.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 
-#include "kv/snapshotter.h"
+#include "node/snapshotter.h"
 
 #include "ccf/ds/logger.h"
 #include "ds/ring_buffer.h"
